Modify incorrect math function.

+++ src/functions/implements/math/abs.c
@@ -85,7 +85,7 @@ rt_function_error_t exec_abs(rt_function_t *f) {
 
   int i; // Iterator
   for (i = 0; i < p->output_size; i++) {
-    y[i] = (float)fabs(x[i]);
+    y[i] = fabsf(x[i]);
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
@@ -98,7 +98,7 @@ rt_function_error_t exec_abs_generic(rt_function_t *f) {
   int i; // Iterator
   for (i = 0; i < p->output_size; i++) {
     float x = p->get_input(p->input, i);
-    p->set_output(p->output, i, (float)fabs(x));
+    p->set_output(p->output, i, fabsf(x));
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
